Refactor onboarding flow into dedicated files
